Multiple SMS count log. Back to all numbers
This fixes the issue when the system sends multiple sms per alarm, now
it counts the amount of sms dividing the total message characters in
groups of 160 characters, that will give us the amount of messages if I
round it up with ceil.

Also I uncommented the log and I ran the all numbers version as seems
to be an stable version.