Station Administration and Cron Close Connection
I have added support to the system in order to bring the list of stations from the admin.redhidro.org service, if it fails, it uses the last copy from that service that is in a file locally at redhidro.org, so in case the communication fails, you may still load the stations thanks to this file, take into account this can be out of date depending on the dashboard administration last changes, as the local file may need manually update.

I also added a mysql close connection on cron at the end of the file just in case.