Optimzation of estado del tiempo
I have created three different store procedures into the database, one for th utp, sedond for the aguas and third for the private stations from utp database, it reduced the load time to 1/3 aprox of the usual time.

I also got rid of a bug that was loading this tab twice, optimizing the load.